Move code from TagExpressionFilter into TagFilter
# Conflicts:
#	documentation/src/docs/asciidoc/release-notes/release-notes-5.1.0-M2.adoc
